requirements.txt
pystray>=0.19
Pillow>=10.0
customtkinter>=5.2
screeninfo>=0.8
